#4209 Variables list can be added to feasibility and amendment forms

+++ mica-webapp/src/main/webapp/assets/js/mica-data-access.js
@@ -82,9 +82,14 @@ class DataAccessService {
       });
   }
 
-  static link(id, type, aId) {
-    let url = '/ws/data-access-request/' + id + '/_link';
+  static linkVariables(id, type, aId) {
+    let url = '/ws/data-access-request/' + id;
     let redirect = '/data-access-form/' + id;
+    if (type && aId) {
+      url = url + '/' + type + '/' + aId;
+      redirect = '/data-access-' + type + '-form/' + aId;
+    }
+    url = url + '/variables';
     axios.put(MicaService.normalizeUrl(url))
       .then(() => {
         //console.dir(response);
@@ -96,9 +101,14 @@ class DataAccessService {
       });
   }
 
-  static unlink(id, type, aId) {
-    let url = '/ws/data-access-request/' + id + '/_link';
+  static unlinkVariables(id, type, aId) {
+    let url = '/ws/data-access-request/' + id;
     let redirect = '/data-access-form/' + id;
+    if (type && aId) {
+      url = url + '/' + type + '/' + aId;
+      redirect = '/data-access-' + type + '-form/' + aId;
+    }
+    url = url + '/variables';
     axios.delete(MicaService.normalizeUrl(url))
       .then(() => {
         //console.dir(response);
